refactor: 로또 개수와 상금 금액 타입 변환

+++ src/main/java/lotto/model/lotto/result/Profit.java
@@ -8,7 +8,7 @@ public class Profit {
     long sumProfitMoney(WinningResult winningResult) {
         init();
         for (Rank rank : winningResult.getWinningCount().keySet()) {
-            number += winningResult.getWinningCount().get(rank) * rank.getValue();
+            number += (long) winningResult.getWinningCount().get(rank) * rank.getValue();
         }
         return number;
     }
